Limit typeRoots to a single node_modules lookup.
Without this, I was using the `yarn add git...`, which meant
tsc was being called in my node_modules/ninjafire/ directory,
and so both ninjafire's node_modules and my other node_modules
were being scanned by tsc.

Specifically a `beforeEach` from in my jest/index.d.ts clashed
with something already in ninjafire's node_modules.

Limiting tsc to just ninjafire's node_modules fixed this for me.

However, it also caused the `console.error` line to fail due to
console not being defined. I'm not entirely sure why this worked
before, but it's the only console.error in the codebase, so I
assume changing it to a `log` call is okay.