Move to a single top-level tsconfig.
My guess is due to the TypeScript upgrade? but 'yarn test' was
failing to have ts-node transpile the TS. It was complaining about
a lack of the Promise type, but both src/tsconfig and test/tsconfig
used lib es016, so that should not have been an issue.

Moving to a single top-level tsconfig fixed it.